requirements.txt
# Core packages
selenium>=4.10.0
beautifulsoup4==4.12.2
pandas>=2.0.3
requests>=2.32.3

# WebDriver management
webdriver-manager>=4.0.0

# Data processing
lxml==4.9.3
html5lib==1.1
regex==2023.6.3

# Utilities
tqdm==4.65.0
urllib3==2.0.3
argparse==1.4.0

# Output formatting
tabulate==0.9.0
openpyxl==3.1.2
xlsxwriter==3.1.0

# Streamlit UI
streamlit==1.31.0
watchdog==3.0.0

# Additional packages
apify-client>=1.6.0
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0
google-api-python-client>=2.169.0
python-dotenv>=1.1.0 